Support multiple key bindings
Use a single constant containing an array of programs to switch to, each
with their own shortcut, window title and command. Create an example
array containing the original Terminator shortcut along with another one
for switching to or launching gworldclock.

Ungrab accelerators first in disable() so that operations are performed
in the reverse order to enable().

Update README to match the new functionality.